+++ flask_unchained/bundles/security/services/user_manager.py
@@ -7,4 +7,5 @@ class UserManager(ModelManager):
     """
     :class:`ModelManager` for the :class:`User` model.
     """
-    model = User
+    class Meta:
+        model = User